refactor(prune_target): extract tests into companion file
`crates/soldr-cli/src/cache_lib/prune_target.rs` was 1035 LOC — over
the soldr-wide 1000-LOC warn threshold enforced by `loc_guard.py`,
with no headroom for future edits. The tail of the file (~450 LOC) is
a self-contained `#[cfg(test)] mod tests {}` block exercising the
prune classifiers; lifting it into `prune_target_tests.rs` and
including it via `#[path]` brings the production file to 583 LOC with
significant headroom.

```
before:  prune_target.rs              1035 LOC
after:   prune_target.rs               583 LOC   ← under warn
         prune_target_tests.rs         456 LOC   ← new, under warn
```

Same pattern soldr already uses elsewhere (`main.rs` → `main_tests.rs`,
`cargo_front_door.rs` → `cargo_front_door_tests.rs`, etc.), now also
applied to the cache_lib submodule. Adds a `cache_lib/README.md` per
the readme_guard hook.

No production-code changes; the test file's `mod tests {}` wrapper
plus inner `use super::*;` is unwrapped so the file IS the module
included via `#[path]`. `system_time_from_unix` is no longer
`#[cfg(test)]`-gated because the entire file is `cfg(test)`-included
at the call site.
