+++ crates/soldr-cli/src/cache_lib/README.md
@@ -1,0 +1,32 @@
+# cache_lib
+
+Cache-management subsystem shared between the `soldr` CLI and the
+soldr-daemon. Owns the on-disk layout under `$SOLDR_CACHE_DIR`, the
+session-state DB, garbage-collection passes, and the per-build `target/`
+prune that ships to zccache via the thin Rust artifact plan.
+
+## Modules
+
+- `mod.rs` — public re-exports + the shared cache-root path helpers.
+- `state_db.rs` — `redb`-backed registry of tracked `target/` dirs and
+  cargo registry-src caches.
+- `auto_gc.rs` / `auto_target_gc.rs` — background GC trigger and the
+  pre/post-build target prune hooks (issue #485).
+- `cargo_global_cache.rs` — `cargo`-native `clean gc` invocation
+  helpers (`$CARGO_HOME` cleanup orchestration).
+- `gc.rs` — soldr-side GC pass orchestrator: locations, summary, purge.
+- `save.rs` — `soldr save` / `soldr load` archive plumbing (mtime-
+  preserving tar.zst bundle of the build cache + source-fingerprint
+  manifest).
+- `prune_target.rs` — per-build `target/` prune (orphan hash-sibling
+  removal + `--keep-latest` aggressive mode). Tests live in
+  `prune_target_tests.rs`, included via `#[path]` so `prune_target.rs`
+  stays under the loc_guard 1K LOC budget.
+
+## Conventions
+
+- The cache root is `$SOLDR_CACHE_DIR` (or `~/.soldr/` when unset). All
+  sub-paths derive from `SoldrPaths` in `crate::core`.
+- Long source files split into `<name>.rs` + `<name>_tests.rs` and
+  include the tests via `#[cfg(test)] #[path = "<name>_tests.rs"] mod
+  tests;` to keep each file under the 1K LOC warn threshold.
